Move to wasm-pack and other WASM build refactoring
A `target` symlink is included until support of workspace is implemented.
