+++ .gitignore
@@ -4,3 +4,14 @@
 */Cargo.lock
 rls/
 .vscode
+
+wasm-pack.log*
+bin/
+pkg/
+
+node_modules
+npm-debug.log*
+yarn-debug.log*
+yarn-error.log*
+.cache
+dist
