+++ playground/src/client.css
@@ -2,12 +2,12 @@
   font-family: 'Material Icons';
   font-style: normal;
   font-weight: 400;
-  src: url(iconfont/MaterialIcons-Regular.eot); /* For IE6-8 */
+  src: url(../node_modules/material-icons/iconfont/MaterialIcons-Regular.eot); /* For IE6-8 */
   src: local('Material Icons'),
     local('MaterialIcons-Regular'),
-    url(iconfont/MaterialIcons-Regular.woff2) format('woff2'),
-    url(iconfont/MaterialIcons-Regular.woff) format('woff'),
-    url(iconfont/MaterialIcons-Regular.ttf) format('truetype');
+    url(../node_modules/material-icons/iconfont/MaterialIcons-Regular.woff2) format('woff2'),
+    url(../node_modules/material-icons/iconfont/MaterialIcons-Regular.woff) format('woff'),
+    url(../node_modules/material-icons/iconfont/MaterialIcons-Regular.ttf) format('truetype');
 }
 
 .material-icons {
