+++ playground/src/index.html
@@ -1,11 +1,10 @@
 <html>
 <head>
 	<title>ratel-wasm</title>
-	<link rel="stylesheet" href="material.indigo-blue.min.css">
-	<link rel="stylesheet" href="client.css">
-	<link rel="shortcut icon" href="favicon.ico">
+	<link rel="stylesheet" href="../node_modules/material-design-lite/material.min.css">
+	<link rel="stylesheet" href="./client.css">
+	<link rel="shortcut icon" href="./favicon.ico">
 	<meta name="viewport" content="width=device-width, initial-scale=1.0">
-	<script defer src="material.min.js"></script>
 </head>
 <body class="ratel">
 	<div class="mdl-grid mdl-grid--no-spacing">
@@ -82,8 +81,7 @@ Array.apply(null, Array(10)).map(function (_, i) {
 		</div>
 	</div>
 
-	<script src="ratel-wasm.js"></script>
-	<script src="clipboard.min.js"></script>
-	<script src="client.js"></script>
+	<script src="../node_modules/material-design-lite/material.min.js"></script>
+	<script src="./client.js"></script>
 </body>
 </html>
