+++ ratel-wasm/Cargo.toml
@@ -3,7 +3,11 @@ name = "ratel-wasm"
 version = "0.0.1"
 authors = ["Matthias Thoemmes <thoemmes@gmail.com>"]
 
+[lib]
+crate-type = ["cdylib"]
+
 [dependencies]
 serde_json = "1.0"
 ratel = { path = "../ratel", version = "0.8.0" }
 ratel-codegen = { path = "../ratel-codegen", version = "0.8.0" }
+wasm-bindgen = "0.2.29"
