--- ratel-wasm/emcc_linker
@@ -1,1 +1,0 @@
-emcc "-s" "WASM=1" "-s" "EXTRA_EXPORTED_RUNTIME_METHODS=['cwrap']" "-s" "ALLOW_MEMORY_GROWTH=1" "-s" "NO_EXIT_RUNTIME=1" $@
